backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
sqlmodel==0.0.14
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.11
httpx==0.26.0
pytest==8.0.0
pytest-asyncio==0.23.5
